refactor: move lockfile types behind facade
